chore: renames api_rest pkg to only rest
