+++ internal/common/api/rest/server.go
@@ -1,4 +1,4 @@
-package api_rest
+package rest
 
 import (
 	"log/slog"
