+++ internal/telemetry/transport/rest/endpoint/v1/create_gps_endpoint.go
@@ -1,7 +1,7 @@
 package v1
 
 import (
-	"github.com/charmingruby/g3/internal/common/api/api_rest"
+	"github.com/charmingruby/g3/internal/common/api/rest"
 	"github.com/charmingruby/g3/internal/common/custom_err"
 	"github.com/charmingruby/g3/internal/telemetry/domain/dto"
 	"github.com/charmingruby/g3/internal/telemetry/transport/rest/endpoint/v1/presenter"
@@ -16,7 +16,7 @@ type CreateGPSRequest struct {
 func (h *Handler) createGPSEndpoint(c *gin.Context) {
 	var req CreateGPSRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
-		api_rest.NewPayloadError(c, err)
+		rest.NewPayloadError(c, err)
 		return
 	}
 
@@ -29,15 +29,15 @@ func (h *Handler) createGPSEndpoint(c *gin.Context) {
 	if err != nil {
 		validationErr, ok := err.(*custom_err.ErrValidation)
 		if ok {
-			api_rest.NewEntityError(c, validationErr)
+			rest.NewEntityError(c, validationErr)
 			return
 		}
 
-		api_rest.NewInternalServerError(c, err)
+		rest.NewInternalServerError(c, err)
 		return
 	}
 
 	mappedGPS := presenter.DomainGPSToHTTP(output.GPS)
 
-	api_rest.NewCreatedResponse(c, "gps", mappedGPS)
+	rest.NewCreatedResponse(c, "gps", mappedGPS)
 }
